refactor(interface): add interface and implements for isp
