+++ src/main/java/racingcar/controller/io/InputHandler.java
@@ -1,0 +1,11 @@
+package racingcar.controller.io;
+
+import camp.nextstep.edu.missionutils.Console;
+import java.util.List;
+
+public interface InputHandler {
+    List<String> getCarInput();
+
+    Integer getIterationInput();
+
+}
