+++ src/main/java/racingcar/model/RacingCar.java
@@ -1,7 +1,6 @@
 package racingcar.model;
 
 import camp.nextstep.edu.missionutils.Randoms;
-import java.util.List;
 
 public class RacingCar {
     private String carName;
