+++ src/main/java/racingcar/view/ConsoleView.java
@@ -1,7 +1,6 @@
 package racingcar.view;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 public class ConsoleView {
     public void printCarInputMessage() {
